+++ components/publish/WriteTestimony.tsx
@@ -8,9 +8,9 @@ import { Maybe } from "../db/common"
 import Input, { InputProps } from "../forms/Input"
 import { useAppDispatch } from "../hooks"
 import { Loading } from "../legislatorSearch"
-import { useFormRedirection } from "./navigation"
+import { useFormRedirection, usePublishState } from "./hooks"
 import * as nav from "./NavigationButtons"
-import { setAttachmentId, setContent, usePublishState } from "./redux"
+import { setAttachmentId, setContent } from "./redux"
 import { StepHeader } from "./StepHeader"
 
 type TabKey = "text" | "import"
