+++ components/publish/hooks/usePublishState.ts
@@ -1,0 +1,9 @@
+import { shallowEqual } from "react-redux"
+import { useAppSelector } from "../../hooks"
+
+export type PublishState = ReturnType<typeof usePublishState>
+export const usePublishState = () =>
+  useAppSelector(
+    ({ publish: { service: edit, ...rest } }) => rest,
+    shallowEqual
+  )
