+++ components/publish/panel/TestimonyFormPanel.tsx
@@ -2,9 +2,8 @@ import { useEffect } from "react"
 import { useAuth } from "../../auth"
 import { Bill } from "../../db"
 import { useAppDispatch } from "../../hooks"
-import { resolveBill } from "../hooks"
+import { resolveBill, usePanelStatus } from "../hooks"
 import { CompleteTestimony, CreateTestimony, SignedOut } from "./ctas"
-import { usePanelStatus } from "./hooks"
 import { ThankYouModal } from "./ThankYouModal"
 import { YourTestimony } from "./YourTestimony"
