+++ components/publish/panel/ThankYouModal.tsx
@@ -2,7 +2,8 @@ import { useCallback, useEffect, useRef, useState } from "react"
 import styled from "styled-components"
 import { Image, Modal } from "../../bootstrap"
 import { useAppDispatch } from "../../hooks"
-import { setShowThankYou, usePublishState } from "../redux"
+import { usePublishState } from "../hooks"
+import { setShowThankYou } from "../redux"
 
 const modalDurationMs = 2000
