+++ components/publish/panel/ctas.tsx
@@ -3,8 +3,7 @@ import { Button, ButtonProps, Image } from "react-bootstrap"
 import styled from "styled-components"
 import { SignInWithModal } from "../../auth"
 import { Wrap } from "../../links"
-import { formUrl } from "../navigation"
-import { usePublishState } from "../redux"
+import { formUrl, usePublishState } from "../hooks"
 
 const Styled = styled.div`
   display: flex;
